# golang/invest-tracker/internal/adapter/controller/notification_controller.go
package controller

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/systentandobr/backend-monorepo/golang/invest-tracker/internal/domain/notification/entity"
	"github.com/systentandobr/backend-monorepo/golang/invest-tracker/pkg/common/logger"
)

// NotificationController handles notification-related HTTP requests
type NotificationController struct {
	logger logger.Logger
}

// NewNotificationController creates a new notification controller
func NewNotificationController(logger logger.Logger) *NotificationController {
	return &NotificationController{
		logger: logger,
	}
}

// RegisterRoutes registers notification routes
func (c *NotificationController) RegisterRoutes(router *gin.RouterGroup) {
	notifications := router.Group("/notifications")
	{
		notifications.GET("", c.GetNotifications)
		notifications.GET("/:id", c.GetNotificationByID)
		notifications.POST("", c.CreateNotification)
		notifications.PUT("/:id", c.UpdateNotification)
		notifications.DELETE("/:id", c.DeleteNotification)
		notifications.PUT("/:id/read", c.MarkAsRead)
		notifications.PUT("/:id/unread", c.MarkAsUnread)
		notifications.GET("/user/:user_id", c.GetUserNotifications)
	}
}

// GetNotifications godoc
// @Summary Get all notifications
// @Description Get a list of all notifications
// @Tags notifications
// @Accept json
// @Produce json
// @Param page query int false "Page number" default(1)
// @Param limit query int false "Items per page" default(10)
// @Param type query string false "Notification type" Enums(alert, reminder, update, system)
// @Param status query string false "Notification status" Enums(read, unread, all)
// @Success 200 {object} map[string]interface{} "List of notifications"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /notifications [get]
func (c *NotificationController) GetNotifications(ctx *gin.Context) {
	page := ctx.DefaultQuery("page", "1")
	limit := ctx.DefaultQuery("limit", "10")
	notificationType := ctx.Query("type")
	status := ctx.Query("status")
	// TODO: Implement get notifications logic
	ctx.JSON(http.StatusOK, gin.H{
		"message": "Get notifications endpoint",
		"page":    page,
		"limit":   limit,
		"type":    notificationType,
		"status":  status,
	})
}

// GetNotificationByID godoc
// @Summary Get notification by ID
// @Description Get a specific notification by its ID
// @Tags notifications
// @Accept json
// @Produce json
// @Param id path string true "Notification ID"
// @Success 200 {object} entity.Notification "Notification details"
// @Failure 404 {object} map[string]interface{} "Notification not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /notifications/{id} [get]
func (c *NotificationController) GetNotificationByID(ctx *gin.Context) {
	id := ctx.Param("id")
	// TODO: Implement get notification by ID logic
	ctx.JSON(http.StatusOK, gin.H{
		"message": "Get notification by ID endpoint",
		"id":      id,
	})
}

// CreateNotification godoc
// @Summary Create new notification
// @Description Create a new notification
// @Tags notifications
// @Accept json
// @Produce json
// @Param notification body entity.Notification true "Notification data"
// @Success 201 {object} entity.Notification "Created notification"
// @Failure 400 {object} map[string]interface{} "Bad request"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /notifications [post]
func (c *NotificationController) CreateNotification(ctx *gin.Context) {
	var notification entity.Notification
	if err := ctx.ShouldBindJSON(&notification); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	// TODO: Implement create notification logic
	ctx.JSON(http.StatusCreated, gin.H{
		"message":      "Create notification endpoint",
		"notification": notification,
	})
}

// UpdateNotification godoc
// @Summary Update notification
// @Description Update an existing notification
// @Tags notifications
// @Accept json
// @Produce json
// @Param id path string true "Notification ID"
// @Param notification body entity.Notification true "Notification data"
// @Success 200 {object} entity.Notification "Updated notification"
// @Failure 400 {object} map[string]interface{} "Bad request"
// @Failure 404 {object} map[string]interface{} "Notification not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /notifications/{id} [put]
func (c *NotificationController) UpdateNotification(ctx *gin.Context) {
	id := ctx.Param("id")
	var notification entity.Notification
	if err := ctx.ShouldBindJSON(&notification); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}
	// TODO: Implement update notification logic
	ctx.JSON(http.StatusOK, gin.H{
		"message":      "Update notification endpoint",
		"id":           id,
		"notification": notification,
	})
}

// DeleteNotification godoc
// @Summary Delete notification
// @Description Delete a notification by ID
// @Tags notifications
// @Accept json
// @Produce json
// @Param id path string true "Notification ID"
// @Success 204 "Notification deleted successfully"
// @Failure 404 {object} map[string]interface{} "Notification not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /notifications/{id} [delete]
func (c *NotificationController) DeleteNotification(ctx *gin.Context) {
	id := ctx.Param("id")
	// TODO: Implement delete notification logic
	ctx.JSON(http.StatusNoContent, gin.H{
		"message": "Delete notification endpoint",
		"id":      id,
	})
}

// MarkAsRead godoc
// @Summary Mark notification as read
// @Description Mark a notification as read
// @Tags notifications
// @Accept json
// @Produce json
// @Param id path string true "Notification ID"
// @Success 200 {object} map[string]interface{} "Notification marked as read"
// @Failure 404 {object} map[string]interface{} "Notification not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /notifications/{id}/read [put]
func (c *NotificationController) MarkAsRead(ctx *gin.Context) {
	id := ctx.Param("id")
	// TODO: Implement mark as read logic
	ctx.JSON(http.StatusOK, gin.H{
		"message": "Mark notification as read endpoint",
		"id":      id,
	})
}

// MarkAsUnread godoc
// @Summary Mark notification as unread
// @Description Mark a notification as unread
// @Tags notifications
// @Accept json
// @Produce json
// @Param id path string true "Notification ID"
// @Success 200 {object} map[string]interface{} "Notification marked as unread"
// @Failure 404 {object} map[string]interface{} "Notification not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /notifications/{id}/unread [put]
func (c *NotificationController) MarkAsUnread(ctx *gin.Context) {
	id := ctx.Param("id")
	// TODO: Implement mark as unread logic
	ctx.JSON(http.StatusOK, gin.H{
		"message": "Mark notification as unread endpoint",
		"id":      id,
	})
}

// GetUserNotifications godoc
// @Summary Get user notifications
// @Description Get notifications for a specific user
// @Tags notifications
// @Accept json
// @Produce json
// @Param user_id path string true "User ID"
// @Param page query int false "Page number" default(1)
// @Param limit query int false "Items per page" default(10)
// @Param status query string false "Notification status" Enums(read, unread, all)
// @Success 200 {object} map[string]interface{} "User notifications"
// @Failure 404 {object} map[string]interface{} "User not found"
// @Failure 500 {object} map[string]interface{} "Internal server error"
// @Router /notifications/user/{user_id} [get]
func (c *NotificationController) GetUserNotifications(ctx *gin.Context) {
	userID := ctx.Param("user_id")
	page := ctx.DefaultQuery("page", "1")
	limit := ctx.DefaultQuery("limit", "10")
	status := ctx.Query("status")
	// TODO: Implement get user notifications logic
	ctx.JSON(http.StatusOK, gin.H{
		"message": "Get user notifications endpoint",
		"user_id": userID,
		"page":    page,
		"limit":   limit,
		"status":  status,
	})
}